fixes to make slide-down animation work
the modal was cutting the slide-down animation short. so now they are synced so that the action sheet slides down, and then the modal hides itself. A bit of hackery was necessary as the `onDismiss` callback of `Modal` isn't working in RN22. perhaps the `onRequestClose` callback `Modal` has in RN23 will come to the rescue, but it I did manually via `setTimeout` for now to guarantee it works across RN versions.